switch to an ArrayOp
