+++ vortex-array/src/arrays/varbinview/compute/mod.rs
@@ -7,7 +7,6 @@ mod is_constant;
 mod is_sorted;
 mod mask;
 mod min_max;
-mod optimize;
 mod take;
 
 #[cfg(test)]
