+++ vortex-array/src/vtable/operations.rs
@@ -28,4 +28,25 @@ pub trait OperationsVTable<V: VTable> {
     /// Bounds-checking has already been performed by the time this function is called,
     /// and the index is guaranteed to be non-null.
     fn scalar_at(array: &V::Array, index: usize) -> VortexResult<Scalar>;
+
+    /// Return an optimized copy of an Array, with any unreferenced data blocks freed and
+    /// any extraneous information removed.
+    ///
+    /// Many simple contiguous array types do not benefit from this operation, but it is
+    /// especially useful for variable-length types which keep a variable number of buffers
+    /// that can be dereferenced by other internal data structures via simple zero-copy
+    /// filter and take operations. After multiple operations are applied to these arrays, it is
+    /// common for the majority of owned buffer data to no longer logically be referenced.
+    ///
+    /// This operation can be called to return a new copy of an array with the same encoding,
+    /// but with all unreferenced data unlinked.
+    ///
+    /// ## Default behavior
+    ///
+    /// For most arrays that do not contain variable buffer counts, such as the canonical
+    /// arrays, the default implementation will not attempt to perform compaction and instead
+    /// return the original array.
+    fn optimize(array: &V::Array) -> VortexResult<ArrayRef> {
+        Ok(array.to_array())
+    }
 }
